Refactor ``get_BZ`` and implement it as ``BZ`` class
The previous `get_BZ` function was refactored and a class was added to represent a BZ. The faces, triangles, hull and delaunaycan be accessed as properties. This is to avoid that the values are changed by the users. Moreover, a new method was added to check if a point is within the BZ.